refactor(proxy): extract ProxyService into lib modules (B7)
- Add cache, pipeline, proxy_service, server, http_types modules
- Move metrics and tls from binary into lib
- Slim main.rs from ~1400 to ~295 lines (thin entrypoint)
- Re-export public API from lib.rs
- Update BLOCKERS, roadmap, and README
