+++ proxy/src/http_types.rs
@@ -1,0 +1,5 @@
+//! Shared HTTP body type alias.
+
+use bytes::Bytes;
+
+pub type Body = http_body_util::Full<Bytes>;
